Use dependencies for checking for Docker
If we rewrite the task Bad Things happen (like a Rake::FileTask turning
into a Rake::Task).  Instead we can use a dependency because that runs
before anything the task does that would require Docker.

PS: rake-hooks is a pointless library.  Rake already has a way to run
things before a task, a prerequisite.  Rake already has a way to run
things after a task, make the after-task have the before-task a
prerequisite.

With some organization of tasks the before/after isn't an issue, it just
falls out with how you hook up your dependencies.