Use #import (as Jim Weirich intended)
Rake has the #import method to safely import Rakefiles into the current
rake application.  Using load or require may cause rakefiles to appear
in the wrong place or require cumbersome setup.

In this case the cumbersome setup was the `cd "mruby"` that required
adjustment for the correct path in the release and package tasks.