Remove too-late compiler check
This check's dependency is :all which runs the compilers so it can never
check what it needs to.

Unfortunately there's no place we can move this compiler check so that
it will work.  Having `task all: :compiler_check` won't work either
because the all task depends on the .o files which run the compiler
check.

I haven't found any tasks in mruby that we can hook into to run such
checks.  I looked extensively to find a place to hook-in for
cross-compiling third party libraries (like libcurl) and such a hook
doesn't exist.

In order to restore this functionality mruby-cli must not import mruby's
Rakefile.  Then we can have tasks that run before mruby is built (like
cross-compiling libraries or checking the cross-compiler environment).